% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaststats-package.R
\docType{package}
\name{blaststats-package}
\alias{blaststats}
\alias{blaststats-package}
\title{blaststats: What the Package Does (Replication of Altschul and Gish 1996 Local Alignment Statistics)}
\description{
Reproduces and explicates the computations and theory behind the p-values and e-values used in BLAST searches.
}
\author{
\strong{Maintainer}: Nathan Brouwer \email{brouwern@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
