% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\docType{data}
\name{table1}
\alias{table1}
\title{Dataset helpfile header . . .}
\format{
A data frame with 12 rows and 7 column(s)
\describe{
\item{mn}{Describe column mn here . . .}
\item{l}{Describe column l here . . .}
\item{ln.nm}{Describe column ln.nm here . . .}
\item{ln.n.m.}{Describe column ln.n.m. here . . .}
\item{u}{Describe column u here . . .}
\item{lambda}{Describe column lambda here . . .}
\item{K}{Describe column K here . . .}
}
}
\source{
\url{http://www.whereyougotthedata.com/}
}
\usage{
table1
}
\description{
Short description of data . . .
}
\keyword{datasets}
